backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.1.3
python-multipart
pydantic[email]
pydantic-settings
google-generativeai
httpx
python-dotenv
Pillow
pytest
